chore: 테스트 패키지 파일 구조 변경
