+++ src/test/java/com/gloddy/server/acceptance/reliability/UpdateReliabilityByEstimateTest.java
@@ -1,14 +1,23 @@
-package com.gloddy.server.reliability;
+package com.gloddy.server.acceptance.reliability;
 
+import com.gloddy.server.auth.entity.User;
 import com.gloddy.server.common.reliability.ReliabilityApiTest;
 import com.gloddy.server.core.event.reliability.ReliabilityScoreUpdateEvent;
-import com.gloddy.server.group.dto.GroupRequest;
+import com.gloddy.server.estimate.dto.EstimateRequest;
+import com.gloddy.server.estimate.entity.embedded.PraiseValue;
+import com.gloddy.server.estimate.service.mate.MateSaveService;
+import com.gloddy.server.estimate.service.praise.PraiseService;
+import com.gloddy.server.group.entity.Group;
+import com.gloddy.server.group.entity.UserGroup;
+import com.gloddy.server.group.service.UserGroupUpdateService;
 import com.gloddy.server.reliability.entity.Reliability;
 import com.gloddy.server.reliability.entity.vo.ReliabilityLevel;
 import com.gloddy.server.reliability.entity.vo.ScorePlusType;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.Commit;
 import org.springframework.test.context.event.ApplicationEvents;
@@ -18,10 +27,18 @@ import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.transaction.annotation.Transactional;
 
-import static org.assertj.core.api.Assertions.*;
 
 @RecordApplicationEvents
-public class UpdateReliabilityByCreateGroupTest extends ReliabilityApiTest {
+public class UpdateReliabilityByEstimateTest extends ReliabilityApiTest {
+
+    @MockBean
+    private PraiseService praiseService;
+
+    @MockBean
+    private MateSaveService mateSaveService;
+
+    @MockBean
+    private UserGroupUpdateService userGroupUpdateService;
 
     @Autowired
     private ApplicationEvents events;
@@ -29,28 +46,26 @@ public class UpdateReliabilityByCreateGroupTest extends ReliabilityApiTest {
     @Test
     @Transactional
     @Commit
-    @DisplayName("그룹 생성 시 신뢰도 점수 업데이트 테스트")
-    void successUpdateReliabilityByCreateGroup() throws Exception {
+    @DisplayName("평가 참여 신뢰도 점수 업데이트 테스트")
+    void successUpdateReliabilityByEstimate() throws Exception {
         // given
-        GroupRequest.Create request = createGroupCreateRequest();
+        User estimateUser = user;
+        User receivePraiseUser = createUser();
+        Group group = createGroup();
+        UserGroup userGroup = createUserGroup(estimateUser, group);
+        EstimateRequest request = createEstimateRequest(receivePraiseUser, PraiseValue.KIND);
 
         // when
-        String url = "/api/v1/group-create";
-        ResultActions result = mockMvc.perform(
-                MockMvcRequestBuilders.post(url)
-                .header("X-AUTH-TOKEN", accessToken)
+        String url = "/api/v1/groups/" + group.getId() + "/estimate";
+        ResultActions result = mockMvc.perform(MockMvcRequestBuilders.post(url)
                 .contentType(MediaType.APPLICATION_JSON)
+                .header("X-AUTH-TOKEN", accessToken)
                 .content(objectMapper.writeValueAsString(request))
         );
 
         // then
-        Reliability reliability = reliabilityQueryHandler.findByUserId(user.getId());
-        assertThat(reliability).isNotNull();
-        assertThat(reliability.getScore()).isEqualTo(0);
-        assertThat(reliability.getLevel()).isEqualTo(ReliabilityLevel.HOOD);
-
         long eventCount = events.stream(ReliabilityScoreUpdateEvent.class).count();
-        assertThat(eventCount).isEqualTo(1);
+        Assertions.assertThat(eventCount).isEqualTo(1);
     }
 
     @AfterTransaction
@@ -59,8 +74,8 @@ public class UpdateReliabilityByCreateGroupTest extends ReliabilityApiTest {
     void afterEvent() {
         Reliability reliability = reliabilityQueryHandler.findByUserId(user.getId());
 
-        assertThat(reliability.getScore()).isEqualTo(ScorePlusType.Created_Group.getScore());
-        assertThat(reliability.getLevel()).isEqualTo(ReliabilityLevel.HOOD);
+        Assertions.assertThat(reliability.getScore()).isEqualTo(ScorePlusType.Estimated.getScore());
+        Assertions.assertThat(reliability.getLevel()).isEqualTo(ReliabilityLevel.HOOD);
 
         userGroupJpaRepository.deleteAll();
         reliabilityRepository.deleteAll();
