+++ src/test/java/com/gloddy/server/unit/reliability/ReliabilityServiceUnitTest.java
@@ -1,4 +1,4 @@
-package com.gloddy.server.reliability;
+package com.gloddy.server.unit.reliability;
 
 import com.gloddy.server.auth.entity.User;
 import com.gloddy.server.reliability.entity.Reliability;
